requirements.txt
Pillow ~= 9.2.0
pyinstaller ~= 5.5
PySide6 ~= 6.4.0
texture2ddecoder ~= 1.0.4
